Chapter reshuffle
As vimtex had a rough time finding the main.tex file for building the
file, I decided to change all the 'main.tex' files in the Chapter
folders into 'chapters.tex'.
I also got rid of the 'STA' folder as I did not need it anymore and
moved the content of said folder into 'eSTA'.