Evaluated robustness with respect to time
I followed the same steps I used in the other type of noise.
Unfortunately in this case the offset in the time variable produces some
numerical instability due to the fact that the polynomial blows up for
t>tf as I did not set up the polynomial to be 0 for t<0 and t>tf.

I will change the source code if needed