Big reshuffle
Created new folders to store the relevant pieces of code to make it more
manageable.
I split the code between the `Fidelities` folder and the `Noise` one.
Also updated the `README.md` file to reflect the changes.
Now I am ready to calculate the effects of the white noise on the eSTA.