All the plots in the notebook
Managed to get the majority of the plots in the notebook.
In particular, the ones that are missing are the ones that show the
evolution of the features as time progresseses.
Next I will add more checks in the `calculations.jl` file to make sure
that no extra data is being calculated.