Interactive Plotting
Finished the pluto notebook to show the plots. There are some niceties I
used to make the whole thing a bit more interactive.
Also, the data are now stored in `data/whole_data.dat`.
Finally, I decided not to add any checks for duplicate as I do not need
it at the moment.