New Combined plots
Printed the plot as pdf files as agreed and changed the code
accordingly.
Next I need to evaluate the value `$\alpha$` that describes the
squeezing of the angular momentum along the x-axis.

Important: in order to change the format of the numbers on the axis
with `PGFPlotsX`, one needs to use `yticklabel_style = "/pgf/number
format/fixed"`.